05_deploy_get_stock_data_image_upload_storage/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
yfinance
mpl_finance
pandas
matplotlib
google-cloud-storage